+++ src/src/observations/observations.module.ts
@@ -1,10 +1,9 @@
-import { Module, NestModule, MiddlewareConsumer } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { HasuraModule } from 'src/hasura/hasura.module';
 import { UserModule } from 'src/user/user.module';
 import { HasuraModule as HasuraModuleFromServices } from '../services/hasura/hasura.module';
 import { ObservationsService } from './observations.service';
 import { ObservationsController } from './observations.controller';
-import { CohortMiddleware } from 'src/common/middlewares/cohort.middleware';
 import { Method } from 'src/common/method/method';
 
 @Module({
@@ -12,10 +11,4 @@ import { Method } from 'src/common/method/method';
 	providers: [ObservationsService, Method],
 	controllers: [ObservationsController],
 })
-export class ObservationsModule implements NestModule {
-	configure(consumer: MiddlewareConsumer) {
-		consumer
-			.apply(CohortMiddleware)
-			.forRoutes('/observations/camp-learner-list');
-	}
-}
+export class ObservationsModule {}
